refacto stats layout
