+++ sveltekit/src/lib/components/StatsNetworks.svelte
